Move DNS watcher to pkg/util/servicediscovery/
